Squashed commit of the following:
commit e8f27095950559b56dd6dd8a94d10d5d7f7af056
Author: Paul LOISIL <fur.waz06@gmail.com>
Date:   Sun May 31 19:15:50 2026 +0200

    Improved web dashboard

commit a34f98cd5e0dad13fc1b2c0596cae5091098b792
Author: Paul LOISIL <fur.waz06@gmail.com>
Date:   Sun May 31 19:15:36 2026 +0200

    Fixed wrong walking gait at startup

commit 5101f0a0e51eebaf248bc6eb4aa66acf4c83a41d
Author: Paul LOISIL <fur.waz06@gmail.com>
Date:   Sun May 31 19:15:17 2026 +0200

    Improved logs & added leg override

commit 52dddeac4547c992cc9f6d9daca949733c988c77
Author: Paul LOISIL <fur.waz06@gmail.com>
Date:   Sun May 31 19:15:02 2026 +0200

    Fixed zerodivision errors

commit 3765a340e27c3759bca1dfad64d0cfefc7f46880
Author: Paul LOISIL <fur.waz06@gmail.com>
Date:   Sun May 31 19:14:37 2026 +0200

    Improved logs, fixed unused config constants

commit b2a9077c7989b8f441576a4cc7fb950bf208eab8
Author: Paul LOISIL <fur.waz06@gmail.com>
Date:   Sun May 31 19:13:02 2026 +0200

    Improved eyes movements

commit b5f2820845aca2745371d7d06e529eb045c85482
Author: Paul LOISIL <fur.waz06@gmail.com>
Date:   Sat May 30 18:15:46 2026 +0200

    Added clearer error codes + increased decision freq

commit e213a806b2085b6394d444a8bc590efd7a0f5f61
Author: Paul LOISIL <fur.waz06@gmail.com>
Date:   Fri May 29 18:58:35 2026 +0200

    Fixed readme presentation

commit a762b48703e43e9a5a10e1727daac402e7c08108
Author: Paul LOISIL <fur.waz06@gmail.com>
Date:   Fri May 29 18:52:15 2026 +0200

    Added clearer links to readme

commit a8290fafa26232a2c61ac495da833057ca0af606
Author: Paul LOISIL <fur.waz06@gmail.com>
Date:   Thu May 28 17:47:03 2026 +0200

    Updated PCBs QRCodes

commit 81563ad7d08ea0afbaef477bd3797e53190ab6d2
Author: Paul LOISIL <fur.waz06@gmail.com>
Date:   Wed May 27 11:55:48 2026 +0200

    Moved SG90 pins to motor driver & added main board fuse

+++ Firmware/include/drivers/MotorDriver.hpp
@@ -15,7 +15,7 @@ namespace MotorDriver
      */
     constexpr static Value MS_TO_PWM(float ms)
     {
-        return 4096.f * ms / (1000.f / MOTOR_DRIVER_PWM_FREQUENCY_HZ);
+        return 4096.f * ms / (1000.f / (float)MOTOR_DRIVER_PWM_FREQUENCY_HZ);
     }
 
     /**
@@ -25,7 +25,7 @@ namespace MotorDriver
      */
     constexpr static float PWM_TO_MS(Value pwm)
     {
-        return pwm * (1000.f / MOTOR_DRIVER_PWM_FREQUENCY_HZ) / 4096.f;
+        return pwm * (1000.f / (float)MOTOR_DRIVER_PWM_FREQUENCY_HZ) / 4096.f;
     }
 
     /**
